+++ CHANGELOG.md
@@ -32,6 +32,7 @@
 * MARC Bib - Hide version history icon and settings if audit log feature is disabled. Refs UIIN-3237.
 * Instance: Suppress action menu and disable buttons when click Change log icon. Refs UIIN-3176.
 * Holdings: Suppress action menu and disable buttons when click Change log icon. Refs UIIN-3177.
+* Adapt settings options for using number gernerator. Refs UIIN-2556.
 
 ## [12.0.12](https://github.com/folio-org/ui-inventory/tree/v12.0.12) (2025-01-27)
 [Full Changelog](https://github.com/folio-org/ui-inventory/compare/v12.0.11...v12.0.12)